requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.12.5
python-dotenv==1.2.1

# LLM & Memory
langchain==0.1.16
langgraph==0.0.64
langchain-ollama==0.1.1
langsmith==0.1.97
ollama==0.1.48

# Database
sqlalchemy==2.0.44

# HTTP
httpx==0.28.1
requests==2.32.5

# Utilities
click==8.3.1
typer==0.12.3
pyyaml==6.0.3
